+++ src/_static/setup-and-start-greenbone-community-edition.sh
@@ -19,6 +19,8 @@
 set -e
 
 DOWNLOAD_DIR=$HOME/greenbone-community-container
+RELEASE="22.4"
+
 
 installed() {
     # $1 should be the command to look for. If $2 is set, we have arguments
@@ -42,8 +44,6 @@ installed() {
 
 }
 
-RELEASE="22.4"
-
 installed curl
 installed docker
 installed docker compose
@@ -55,16 +55,16 @@ mkdir -p $DOWNLOAD_DIR && cd $DOWNLOAD_DIR
 echo "Downloading docker-compose file..."
 curl -f -O https://greenbone.github.io/docs/latest/_static/docker-compose.yml
 
-echo "Pulling Greenbone Community Containers $RELEASE"
-docker compose -f $DOWNLOAD_DIR/docker-compose-$RELEASE.yml pull
+echo "Pulling Greenbone Community Containers"
+docker compose -f $DOWNLOAD_DIR/docker-compose.yml pull
 echo
 
-echo "Starting Greenbone Community Containers $RELEASE"
-docker compose -f $DOWNLOAD_DIR/docker-compose-$RELEASE.yml up -d
+echo "Starting Greenbone Community Containers"
+docker compose -f $DOWNLOAD_DIR/docker-compose.yml up -d
 echo
 
 read -s -p "Password for admin user: " password
-docker compose -f $DOWNLOAD_DIR/docker-compose-$RELEASE.yml \
+docker compose -f $DOWNLOAD_DIR/docker-compose.yml \
     exec -u gvmd gvmd gvmd --user=admin --new-password=$password
 
 echo